parsing

+++ app/src/main/java/it/unibo/bbc_smartphone/model/Alert.java
@@ -4,8 +4,8 @@ package it.unibo.bbc_smartphone.model;
  * Created by brando on 15/06/2015.
  */
 public class Alert {
-    private long latitude;
-    private long longitude;
+    private double latitude;
+    private double longitude;
     private String message;
     private int idPlayer;
 
@@ -16,7 +16,7 @@ public class Alert {
         this.idPlayer = idPlayer;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
@@ -24,7 +24,7 @@ public class Alert {
         this.latitude = latitude;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
