+++ app/src/main/java/it/unibo/bbc_smartphone/model/TreasureChest.java
@@ -5,8 +5,8 @@ package it.unibo.bbc_smartphone.model;
  */
 public class TreasureChest {
     private int number;
-    private long latitude;
-    private long longitude;
+    private double latitude;
+    private double longitude;
     private int money;
     public enum State{
         UNVISITED, OPEN, LOCKED_KEY, LOCKED_COOPERATION, FINAL;
@@ -30,7 +30,7 @@ public class TreasureChest {
         this.number = number;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
@@ -38,7 +38,7 @@ public class TreasureChest {
         this.latitude = latitude;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
